# app/api/v1/feed/route.ts
/**
 * @module app/api/v1/feed
 * GET /api/v1/feed — 認証ユーザーのタイムライン取得
 */

import { type NextRequest, NextResponse } from 'next/server'
import { requireBearerUser, apiZodError, apiRateLimited } from '@/lib/api/v1'
import { checkUserRateLimit } from '@/lib/rate-limit'
import { readPaginationQuerySchema } from '@/lib/api/v1/schemas/request'
import { fetchTimeline } from '@/lib/services/feed-service'

export async function GET(request: NextRequest) {
  // 1. Bearer 認証
  const auth = await requireBearerUser(request)
  if (!auth.ok) return auth.response

  // 2. クエリパラメータ検証
  const { searchParams } = request.nextUrl
  const parsed = readPaginationQuerySchema.safeParse({
    cursor: searchParams.get('cursor') ?? undefined,
    limit: searchParams.has('limit') ? Number(searchParams.get('limit')) : undefined,
  })
  if (!parsed.success) return apiZodError(parsed.error)

  // 3. レート制限
  const rl = await checkUserRateLimit(auth.userId, 'timeline')
  if (!rl.success) return apiRateLimited(rl)

  // 4. タイムライン取得
  const result = await fetchTimeline(auth.userId, parsed.data.cursor, parsed.data.limit)

  return NextResponse.json({
    items: result.posts,
    nextCursor: result.nextCursor ?? null,
    isGuest: result.isGuest,
  })
}