# openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Bon_Log Mobile API",
    "version": "1.1.0",
    "description": "盆栽 SNS「Bon_Log」のモバイルアプリ向け API。\n\n## 認証フロー\n\n1. POST /api/v1/auth/login でトークンペアを取得\n2. API リクエストに Authorization: Bearer <accessToken> を付与\n3. 401 AUTH_TOKEN_EXPIRED を受け取ったら POST /api/v1/auth/refresh でペアを更新\n4. 401 AUTH_REFRESH_TOKEN_REUSE_DETECTED はセキュリティ警告画面を表示して再ログインを促す\n\n## レート制限\n\n429 レスポンスには Retry-After ヘッダー（秒）が付与される。自動リトライは禁止。"
  },
  "servers": [
    {
      "url": "https://www.bon-log.com",
      "description": "本番サーバー"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "アクセストークン（JWT HS256, TTL 15 分）。期限切れ時は /api/v1/auth/refresh でトークンペアを更新する。"
      }
    },
    "schemas": {
...
      "ApiErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string",
                "enum": [
                  "AUTH_REQUIRED",
                  "AUTH_INVALID_TOKEN",
                  "AUTH_TOKEN_EXPIRED",
                  "AUTH_INVALID_CREDENTIALS",
                  "AUTH_2FA_REQUIRED",
                  "AUTH_2FA_INVALID_CODE",
                  "AUTH_2FA_TICKET_EXPIRED",
                  "AUTH_REFRESH_TOKEN_INVALID",
                  "AUTH_REFRESH_TOKEN_REUSE_DETECTED",
                  "ACCOUNT_SUSPENDED",
                  "GUEST_NOT_ALLOWED",
                  "EMAIL_NOT_VERIFIED",
                  "VALIDATION_ERROR",
                  "RATE_LIMITED",
                  "NOT_FOUND",
                  "CONFLICT",
                  "INTERNAL_ERROR",
                  "SERVER_MISCONFIGURED"
                ]
              },
              "message": {
                "type": "string"
              },
              "status": {
                "type": "integer"
              }
            },
            "required": [
              "code",
              "message",
              "status"
            ]
          }
        },
        "required": [
          "error"
        ],
        "description": "全エンドポイント共通エラーレスポンス形式。"
      },
...
      "FeedResponse": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "string"
                },
                "content": {
                  "type": "string"
                },
                "createdAt": {
                  "type": "string",
                  "format": "date-time"
                },
                "updatedAt": {
                  "type": "string",
                  "format": "date-time"
                },
                "userId": {
                  "type": "string"
                },
                "user": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string"
                    },
                    "nickname": {
                      "type": "string"
                    },
                    "avatarUrl": {
                      "type": [
                        "string",
                        "null"
                      ]
                    }
                  },
                  "required": [
                    "id",
                    "nickname",
                    "avatarUrl"
                  ]
                },
                "media": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "url": {
                        "type": "string"
                      },
                      "type": {
                        "type": "string"
                      },
                      "sortOrder": {
                        "type": "integer"
                      }
                    },
                    "required": [
                      "url",
                      "type",
                      "sortOrder"
                    ]
                  }
                },
                "genres": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "string"
                      },
                      "name": {
                        "type": "string"
                      },
                      "category": {
                        "type": "string"
                      }
                    },
                    "required": [
                      "id",
                      "name",
                      "category"
                    ]
                  }
                },
                "likeCount": {
                  "type": "integer"
                },
                "commentCount": {
                  "type": "integer"
                },
                "repostCount": {
                  "type": "integer"
                },
                "isLiked": {
                  "type": "boolean"
                },
                "isBookmarked": {
                  "type": "boolean"
                },
                "isReposted": {
                  "type": "boolean"
                },
                "quotePost": {
                  "type": [
                    "object",
                    "null"
                  ],
                  "properties": {
                    "id": {
                      "type": "string"
                    },
                    "content": {
                      "type": "string"
                    },
                    "user": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string"
                        },
                        "nickname": {
                          "type": "string"
                        },
                        "avatarUrl": {
                          "type": [
                            "string",
                            "null"
                          ]
                        }
                      },
                      "required": [
                        "id",
                        "nickname",
                        "avatarUrl"
                      ]
                    },
                    "media": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "url": {
                            "type": "string"
                          },
                          "type": {
                            "type": "string"
                          },
                          "sortOrder": {
                            "type": "integer"
                          }
                        },
                        "required": [
                          "url",
                          "type",
                          "sortOrder"
                        ]
                      }
                    }
                  },
                  "required": [
                    "id",
                    "content",
                    "user"
                  ]
                },
                "repostPost": {
                  "type": [
                    "object",
                    "null"
                  ],
                  "properties": {
                    "id": {
                      "type": "string"
                    },
                    "content": {
                      "type": "string"
                    },
                    "user": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string"
                        },
                        "nickname": {
                          "type": "string"
                        },
                        "avatarUrl": {
                          "type": [
                            "string",
                            "null"
                          ]
                        }
                      },
                      "required": [
                        "id",
                        "nickname",
                        "avatarUrl"
                      ]
                    },
                    "media": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "url": {
                            "type": "string"
                          },
                          "type": {
                            "type": "string"
                          },
                          "sortOrder": {
                            "type": "integer"
                          }
                        },
                        "required": [
                          "url",
                          "type",
                          "sortOrder"
                        ]
                      }
                    }
                  },
                  "required": [
                    "id",
                    "content",
                    "user"
                  ]
                },
                "poll": {}
              },
              "required": [
                "id",
                "content",
                "createdAt",
                "updatedAt",
                "userId",
                "user",
                "media",
                "genres",
                "likeCount",
                "commentCount",
                "repostCount",
                "isLiked",
                "isBookmarked",
                "isReposted",
                "quotePost",
                "repostPost"
              ]
            }
          },
          "nextCursor": {
            "type": [
              "string",
              "null"
            ]
          },
          "isGuest": {
            "type": "boolean"
          }
        },
        "required": [
          "items",
          "nextCursor",
          "isGuest"
        ],
        "description": "タイムライン取得レスポンス。"
      },
...
      "CommentsListResponse": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "string"
                },
                "postId": {
                  "type": "string"
                },
                "userId": {
                  "type": "string"
                },
                "parentId": {
                  "type": [
                    "string",
                    "null"
                  ]
                },
                "content": {
                  "type": "string"
                },
                "createdAt": {
                  "type": "string",
                  "format": "date-time"
                },
                "updatedAt": {
                  "type": "string",
                  "format": "date-time"
                },
                "isDeleted": {
                  "type": "boolean"
                },
                "isBlockedUser": {
                  "type": "boolean"
                },
                "likeCount": {
                  "type": "integer"
                },
                "replyCount": {
                  "type": "integer"
                },
                "isLiked": {
                  "type": "boolean"
                },
                "user": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string"
                    },
                    "nickname": {
                      "type": "string"
                    },
                    "avatarUrl": {
                      "type": [
                        "string",
                        "null"
                      ]
                    }
                  },
                  "required": [
                    "id",
                    "nickname",
                    "avatarUrl"
                  ]
                },
                "media": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "url": {
                        "type": "string"
                      },
                      "type": {
                        "type": "string"
                      },
                      "sortOrder": {
                        "type": "integer"
                      }
                    },
                    "required": [
                      "url",
                      "type",
                      "sortOrder"
                    ]
                  }
                }
              },
              "required": [
                "id",
                "postId",
                "userId",
                "parentId",
                "content",
                "createdAt",
                "updatedAt",
                "isDeleted",
                "isBlockedUser",
                "likeCount",
                "replyCount",
                "isLiked",
                "user",
                "media"
              ]
            }
          },
          "nextCursor": {
            "type": [
              "string",
              "null"
            ]
          }
        },
        "required": [
          "items",
          "nextCursor"
        ],
        "description": "コメント一覧取得レスポンス。"
      },
...
      "SearchUsersResponse": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "string"
                },
                "nickname": {
                  "type": "string"
                },
                "avatarUrl": {
                  "type": [
                    "string",
                    "null"
                  ]
                },
                "bio": {
                  "type": [
                    "string",
                    "null"
                  ]
                },
                "followersCount": {
                  "type": "integer"
                },
                "followingCount": {
                  "type": "integer"
                }
              },
              "required": [
                "id",
                "nickname",
                "avatarUrl",
                "bio",
                "followersCount",
                "followingCount"
              ]
            }
          },
          "nextCursor": {
            "type": [
              "string",
              "null"
            ]
          }
        },
        "required": [
          "items",
          "nextCursor"
        ],
        "description": "ユーザー検索結果レスポンス。"
      },
...
    },
...
  },
  "paths": {
...
    "/api/v1/feed": {
      "get": {
        "tags": [
          "feed"
        ],
        "summary": "タイムライン取得",
        "description": "認証ユーザーのタイムライン（フォロー中ユーザーの投稿）をカーソルページネーションで返す。\n\nゲストユーザーの場合は公開投稿の直近 N 件のみ返し nextCursor は null になる。\n\nブロック・ミュート・停止済み著者の投稿は除外される。",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "description": "カーソル（前回レスポンスの nextCursor 値）"
            },
            "required": false,
            "description": "カーソル（前回レスポンスの nextCursor 値）",
            "name": "cursor",
            "in": "query"
          },
          {
            "schema": {
              "type": "integer",
              "description": "1 回の取得上限件数"
            },
            "required": false,
            "description": "1 回の取得上限件数",
            "name": "limit",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "タイムライン取得成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FeedResponse"
                }
              }
            }
          },
          "400": {
            "description": "バリデーションエラー (VALIDATION_ERROR)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Bearer トークンなし (AUTH_REQUIRED) または期限切れ (AUTH_TOKEN_EXPIRED)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "アカウント停止 (ACCOUNT_SUSPENDED)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "レート制限超過。Retry-After ヘッダー（秒）が返却される。自動リトライ禁止。",
            "headers": {
              "Retry-After": {
                "description": "次のリクエストまでの待機秒数",
                "schema": {
                  "type": "integer"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/posts/{id}/comments": {
      "get": {
        "tags": [
          "posts"
        ],
        "summary": "コメント一覧取得",
        "description": "指定投稿のトップレベルコメント一覧をカーソルページネーションで返す。\n\nブロック済みユーザーのコメントは isBlockedUser: true として含まれる（非表示はクライアント側で判断）。",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "description": "投稿 ID"
            },
            "required": true,
            "description": "投稿 ID",
            "name": "id",
            "in": "path"
          },
          {
            "schema": {
              "type": "string",
              "description": "カーソル"
            },
            "required": false,
            "description": "カーソル",
            "name": "cursor",
            "in": "query"
          },
          {
            "schema": {
              "type": "integer",
              "description": "取得上限件数"
            },
            "required": false,
            "description": "取得上限件数",
            "name": "limit",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "コメント一覧",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommentsListResponse"
                }
              }
            }
          },
          "400": {
            "description": "バリデーションエラー (VALIDATION_ERROR)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Bearer トークンなし (AUTH_REQUIRED) または期限切れ (AUTH_TOKEN_EXPIRED)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "アカウント停止 (ACCOUNT_SUSPENDED)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "投稿が存在しないか閲覧権限なし (NOT_FOUND)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "レート制限超過。Retry-After ヘッダー（秒）が返却される。自動リトライ禁止。",
            "headers": {
              "Retry-After": {
                "description": "次のリクエストまでの待機秒数",
                "schema": {
                  "type": "integer"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/search/posts": {
      "get": {
        "tags": [
          "search"
        ],
        "summary": "投稿検索",
        "description": "キーワードで投稿を全文検索する。\n\n2 文字以上: pg_bigm trigram 検索。1 文字: LIKE フォールバック。\nブロック・ミュート・非公開・停止ユーザーの投稿は除外される。",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "description": "検索キーワード"
            },
            "required": false,
            "description": "検索キーワード",
            "name": "q",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "description": "カーソル"
            },
            "required": false,
            "description": "カーソル",
            "name": "cursor",
            "in": "query"
          },
          {
            "schema": {
              "type": "integer",
              "description": "取得上限件数"
            },
            "required": false,
            "description": "取得上限件数",
            "name": "limit",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "投稿検索結果",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchPostsResponse"
                }
              }
            }
          },
          "400": {
            "description": "バリデーションエラー (VALIDATION_ERROR)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Bearer トークンなし (AUTH_REQUIRED) または期限切れ (AUTH_TOKEN_EXPIRED)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "アカウント停止 (ACCOUNT_SUSPENDED)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "レート制限超過。Retry-After ヘッダー（秒）が返却される。自動リトライ禁止。",
            "headers": {
              "Retry-After": {
                "description": "次のリクエストまでの待機秒数",
                "schema": {
                  "type": "integer"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/search/users": {
      "get": {
        "tags": [
          "search"
        ],
        "summary": "ユーザー検索",
        "description": "ニックネームでユーザーを検索する。\n\nブロック・ミュート・停止済みユーザーは除外される。\nゲストアカウントは結果に含まれない。",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "description": "検索キーワード"
            },
            "required": false,
            "description": "検索キーワード",
            "name": "q",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "description": "カーソル"
            },
            "required": false,
            "description": "カーソル",
            "name": "cursor",
            "in": "query"
          },
          {
            "schema": {
              "type": "integer",
              "description": "取得上限件数"
            },
            "required": false,
            "description": "取得上限件数",
            "name": "limit",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "ユーザー検索結果",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchUsersResponse"
                }
              }
            }
          },
          "400": {
            "description": "バリデーションエラー (VALIDATION_ERROR)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Bearer トークンなし (AUTH_REQUIRED) または期限切れ (AUTH_TOKEN_EXPIRED)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "アカウント停止 (ACCOUNT_SUSPENDED)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "レート制限超過。Retry-After ヘッダー（秒）が返却される。自動リトライ禁止。",
            "headers": {
              "Retry-After": {
                "description": "次のリクエストまでの待機秒数",
                "schema": {
                  "type": "integer"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/notifications": {
      "get": {
        "tags": [
          "notifications"
        ],
        "summary": "通知一覧取得",
        "description": "認証ユーザーの通知をカーソルページネーションで返す。\n\nゲストユーザーは利用不可（403 GUEST_NOT_ALLOWED）。",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "description": "カーソル"
            },
            "required": false,
            "description": "カーソル",
            "name": "cursor",
            "in": "query"
          },
          {
            "schema": {
              "type": "integer",
              "description": "取得上限件数"
            },
            "required": false,
            "description": "取得上限件数",
            "name": "limit",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "通知一覧",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotificationsListResponse"
                }
              }
            }
          },
          "400": {
            "description": "バリデーションエラー (VALIDATION_ERROR)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Bearer トークンなし (AUTH_REQUIRED) または期限切れ (AUTH_TOKEN_EXPIRED)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "アカウント停止 (ACCOUNT_SUSPENDED) またはゲスト不可 (GUEST_NOT_ALLOWED)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "レート制限超過。Retry-After ヘッダー（秒）が返却される。自動リトライ禁止。",
            "headers": {
              "Retry-After": {
                "description": "次のリクエストまでの待機秒数",
                "schema": {
                  "type": "integer"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}